pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref>=3.8.1",
    "django>=5.1.15",
    "django-nested-admin>=4.1.6",
    "gunicorn>=23.0.0",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlparse>=0.5.5",
    "typing-extensions>=4.13.2",
]

[project.optional-dependencies]
dev = [
    "black>=26.1.0",
    "isort>=8.0.0",
    "flake8>=7.3.0",
    "flake8-docstrings>=1.7.0",
    "flake8-bugbear>=25.11.29",
    "pylint>=4.0.5",
    "mypy>=1.19.1",
    "django-stubs>=5.2.9",
    "bandit[toml]>=1.9.3",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "mysite.settings"

[tool.pylint.main]
load-plugins = ["pylint_django"]
django-settings-module = "mysite.settings"
ignore = ["migrations", ".venv", "venv"]

[tool.pylint.format]
max-line-length = 88

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "R0903",  # too-few-public-methods
]

[tool.bandit]
exclude_dirs = ["/tests", "*/migrations/*", ".venv", "venv"]
skips = ["B101", "B601"]

[tool.bandit.assert_used]
skips = ["*/test_*.py", "*/tests.py"]


[tool.uv]
constraint-dependencies = [
    "virtualenv>=20.36.1",
]

[tool.setuptools]
packages = ["apps", "mysite"]

[tool.setuptools.package-data]
"*" = ["*.html", "*.css", "*.js", "*.json", "*.txt", "*.po", "*.mo"]
